bd-h3dib.2: move storage conformance suite to public backend/conformance
Relocate internal/storage/conformance to backend/conformance so external
Go modules can import and run the suite against out-of-tree storage
backends (the conformance-gated external-backend path promised on #4415).
Pure move: package name, entry points, and assertions are unchanged; the
seven in-tree test importers are rewritten and path references in doc
comments updated. The package doc now states the external usage and the
experimental stability terms.
