+++ internal/storage/dolt/issue_operations_staging_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/steveyegge/beads/internal/storage/conformance"
+	"github.com/steveyegge/beads/backend/conformance"
 	"github.com/steveyegge/beads/internal/types"
 	publicops "github.com/steveyegge/beads/issueops"
 )
