+++ internal/storage/embeddeddolt/issue_operations_contract_test.go
@@ -6,7 +6,7 @@ import (
 	"context"
 	"testing"
 
-	"github.com/steveyegge/beads/internal/storage/conformance"
+	"github.com/steveyegge/beads/backend/conformance"
 	"github.com/steveyegge/beads/internal/storage/embeddeddolt"
 )
